requirements.txt
Flask==2.3.2
Flask-JSON==0.4.0
kafka-python==2.0.2
keras==2.13.1
matplotlib==3.7.2
numpy==1.24.3
pandas==2.0.3
psycopg2-binary==2.9.6
python-dotenv==1.0.0
scikit-learn==1.3.0
tensorflow==2.13.0
waitress==2.1.2
yfinance==0.2.25